# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kadonation - PHP Developer Challenge API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "L5 Swagger OpenApi dynamic host server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "operationId": "39c74c9badb6c7c74d2612d1c7eeb685",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a product",
                "operationId": "0254f499bf820a11da3ca96eeb4eaf9f",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ProductRequest"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show a product",
                "operationId": "97ac93a39c2d8858ad097f9581ef721b",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "9117ad6b2d52d5052030104f75249f95",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "80822b23004c18331e67dbafa86a1d42",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ProductRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost/api/resource?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost/api/resource?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": "http://localhost/api/resource?page=3"
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost/api/resource?page=4"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost/api/resource"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 4
                    },
                    "name": {
                        "type": "string",
                        "example": "Furnitures"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2021-08-07T05:54:34.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2021-08-07T05:54:34.000000Z"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Cyberpunk 2077"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "sku": {
                        "type": "string",
                        "example": "A0005"
                    },
                    "price": {
                        "type": "number",
                        "example": 59.99
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2021-08-07T05:54:34.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2021-08-07T05:54:34.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "ProductRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Cyberpunk 2077"
                                },
                                "category_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "sku": {
                                    "type": "string",
                                    "example": "A0005"
                                },
                                "price": {
                                    "type": "number",
                                    "example": 59.99
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 20
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}